use shared configuration between proofing job and webhook caller
